Moves RouterSpy to private scope as it is not shared anymore
