+++ QuizEngineTests/FlowTest.swift
@@ -10,7 +10,7 @@ import XCTest
 @testable import QuizEngine
 
 class FlowTest: XCTestCase {
-    let router = RouterSpy()
+    private let router = RouterSpy()
 
     func test_start_withNoQuestions_doesNotRouteToQuestion() {
         makeSUT(questions: []).start()
@@ -124,7 +124,23 @@ class FlowTest: XCTestCase {
     
     // MARK: - Helpers
     
-    func makeSUT(questions: [String], scoring: @escaping ([String: String]) -> Int = { _ in 0 }) -> Flow<String, String, RouterSpy> {
+    private func makeSUT(questions: [String], scoring: @escaping ([String: String]) -> Int = { _ in 0 }) -> Flow<String, String, RouterSpy> {
         return Flow(questions: questions, router: router, scoring: scoring)
+    }
+    
+    private final class RouterSpy: Router {
+        var routedQuestions: [String] = []
+        var routedResults: Result<String, String>? = nil
+        
+        var answerCallback: ((String) -> Void) = { _ in }
+        
+        func routeTo(question: String, answerCallback: @escaping (String) -> Void) {
+            routedQuestions.append(question)
+            self.answerCallback = answerCallback
+        }
+        
+        func routeTo(result: Result<String, String>) {
+            routedResults = result
+        }
     }
 }
